refactor: normalize screens as folders and own plans UI
Every screen is a folder with index.tsx. Move plans-view and
usePlansScreen under plans-screen, and colocate useDashboardData
and useDayTasks with their screens.
