--- components/plans-view/index.ts
@@ -1,4 +1,0 @@
-export * from "./context-view/plans-context-view";
-export * from "./context-view/types";
-export * from "./month-view/plans-month-view";
-export * from "./month-view/types";
