+++ screens/dashboard-screen/index.tsx
@@ -1,7 +1,7 @@
 import { ScrollView } from "@/components/ui/scroll-view";
 import { Box } from "@/components/ui/box";
 import { memo } from "react";
-import { useDashboardData } from "../../hooks/useDashboardData";
+import { useDashboardData } from "./hooks/useDashboardData";
 import { EmptyScreen } from "../../components/empty-screen";
 import { DashboardStats } from "./dashboard-stats";
 import { ContextSections } from "./context-sections";
