+++ screens/intro-screen/index.tsx
@@ -2,16 +2,16 @@ import { Button, ButtonText } from "@/components/ui/button";
 import { Center } from "@/components/ui/center";
 import { Heading } from "@/components/ui/heading";
 import { Box } from "@/components/ui/box";
-import { SafeAreaView } from "../components/common/safe-area-view";
-import { SnowAngel, Decorating, Dog, SkiingSantaSvg } from "../assets/svg";
+import { SafeAreaView } from "../../components/common/safe-area-view";
+import { SnowAngel, Decorating, Dog, SkiingSantaSvg } from "../../assets/svg";
 import Carousel, { Pagination } from "react-native-reanimated-carousel";
 import { useCallback, useRef, useState, ComponentType } from "react";
 import { LayoutChangeEvent } from "react-native";
 import { useTranslation } from "react-i18next";
 import { useNavigation } from "@react-navigation/native";
-import { SCREENS } from "../constants/constants";
+import { SCREENS } from "../../constants/constants";
 import { StackNavigationProp } from "@react-navigation/stack";
-import { RootStackParamList } from "../App";
+import { RootStackParamList } from "../../App";
 import * as Haptics from "expo-haptics";
 import { useSharedValue } from "react-native-reanimated";
 import { SvgProps } from "react-native-svg";
