+++ screens/plans-screen/context-view/content.tsx
